requirements.txt
uvicorn
fastapi
grpcio
protobuf
grpcio-tools
sqlalchemy
web3
asyncpg
